Move fmt test cases.
