+++ contracts/oracle/osmosis/src/price_source.rs
@@ -757,159 +757,6 @@ mod tests {
     use super::*;
 
     #[test]
-    fn display_downtime_detector() {
-        let dd = DowntimeDetector {
-            downtime: Downtime::Duration10m,
-            recovery: 550,
-        };
-        assert_eq!(dd.to_string(), "Duration10m:550")
-    }
-
-    #[test]
-    fn display_fixed_price_source() {
-        let ps = OsmosisPriceSourceChecked::Fixed {
-            price: Decimal::from_ratio(1u128, 2u128),
-        };
-        assert_eq!(ps.to_string(), "fixed:0.5")
-    }
-
-    #[test]
-    fn display_spot_price_source() {
-        let ps = OsmosisPriceSourceChecked::Spot {
-            pool_id: 123,
-        };
-        assert_eq!(ps.to_string(), "spot:123")
-    }
-
-    #[test]
-    fn display_arithmetic_twap_price_source() {
-        let ps = OsmosisPriceSourceChecked::ArithmeticTwap {
-            pool_id: 123,
-            window_size: 300,
-            downtime_detector: None,
-        };
-        assert_eq!(ps.to_string(), "arithmetic_twap:123:300:None");
-
-        let ps = OsmosisPriceSourceChecked::ArithmeticTwap {
-            pool_id: 123,
-            window_size: 300,
-            downtime_detector: Some(DowntimeDetector {
-                downtime: Downtime::Duration30m,
-                recovery: 568,
-            }),
-        };
-        assert_eq!(ps.to_string(), "arithmetic_twap:123:300:Some(Duration30m:568)");
-    }
-
-    #[test]
-    fn display_geometric_twap_price_source() {
-        let ps = OsmosisPriceSourceChecked::GeometricTwap {
-            pool_id: 123,
-            window_size: 300,
-            downtime_detector: None,
-        };
-        assert_eq!(ps.to_string(), "geometric_twap:123:300:None");
-
-        let ps = OsmosisPriceSourceChecked::GeometricTwap {
-            pool_id: 123,
-            window_size: 300,
-            downtime_detector: Some(DowntimeDetector {
-                downtime: Downtime::Duration30m,
-                recovery: 568,
-            }),
-        };
-        assert_eq!(ps.to_string(), "geometric_twap:123:300:Some(Duration30m:568)");
-    }
-
-    #[test]
-    fn display_staked_geometric_twap_price_source() {
-        let ps = OsmosisPriceSourceChecked::StakedGeometricTwap {
-            transitive_denom: "transitive".to_string(),
-            pool_id: 123,
-            window_size: 300,
-            downtime_detector: None,
-        };
-        assert_eq!(ps.to_string(), "staked_geometric_twap:transitive:123:300:None");
-
-        let ps = OsmosisPriceSourceChecked::StakedGeometricTwap {
-            transitive_denom: "transitive".to_string(),
-            pool_id: 123,
-            window_size: 300,
-            downtime_detector: Some(DowntimeDetector {
-                downtime: Downtime::Duration30m,
-                recovery: 568,
-            }),
-        };
-        assert_eq!(
-            ps.to_string(),
-            "staked_geometric_twap:transitive:123:300:Some(Duration30m:568)"
-        );
-    }
-
-    #[test]
-    fn display_xyk_lp_price_source() {
-        let ps = OsmosisPriceSourceChecked::XykLiquidityToken {
-            pool_id: 224,
-        };
-        assert_eq!(ps.to_string(), "xyk_liquidity_token:224")
-    }
-
-    #[test]
-    fn display_pyth_price_source() {
-        let ps = OsmosisPriceSourceChecked::Pyth {
-            contract_addr: Addr::unchecked("osmo12j43nf2f0qumnt2zrrmpvnsqgzndxefujlvr08"),
-            price_feed_id: PriceIdentifier::from_hex(
-                "61226d39beea19d334f17c2febce27e12646d84675924ebb02b9cdaea68727e3",
-            )
-            .unwrap(),
-            max_staleness: 60,
-            denom_decimals: 18,
-        };
-        assert_eq!(
-            ps.to_string(),
-            "pyth:osmo12j43nf2f0qumnt2zrrmpvnsqgzndxefujlvr08:0x61226d39beea19d334f17c2febce27e12646d84675924ebb02b9cdaea68727e3:60:18"
-        )
-    }
-
-    #[test]
-    fn display_lsd_price_source() {
-        let ps = OsmosisPriceSourceChecked::Lsd {
-            transitive_denom: "transitive".to_string(),
-            geometric_twap: GeometricTwap {
-                pool_id: 456,
-                window_size: 380,
-                downtime_detector: None,
-            },
-            redemption_rate: RedemptionRate {
-                contract_addr: Addr::unchecked(
-                    "osmo1zw4fxj4pt0pu0jdd7cs6gecdj3pvfxhhtgkm4w2y44jp60hywzvssud6uc",
-                ),
-                max_staleness: 1234,
-            },
-        };
-        assert_eq!(ps.to_string(), "lsd:transitive:456:380:None:osmo1zw4fxj4pt0pu0jdd7cs6gecdj3pvfxhhtgkm4w2y44jp60hywzvssud6uc:1234");
-
-        let ps = OsmosisPriceSourceChecked::Lsd {
-            transitive_denom: "transitive".to_string(),
-            geometric_twap: GeometricTwap {
-                pool_id: 456,
-                window_size: 380,
-                downtime_detector: Some(DowntimeDetector {
-                    downtime: Downtime::Duration30m,
-                    recovery: 552,
-                }),
-            },
-            redemption_rate: RedemptionRate {
-                contract_addr: Addr::unchecked(
-                    "osmo1zw4fxj4pt0pu0jdd7cs6gecdj3pvfxhhtgkm4w2y44jp60hywzvssud6uc",
-                ),
-                max_staleness: 1234,
-            },
-        };
-        assert_eq!(ps.to_string(), "lsd:transitive:456:380:Some(Duration30m:552):osmo1zw4fxj4pt0pu0jdd7cs6gecdj3pvfxhhtgkm4w2y44jp60hywzvssud6uc:1234");
-    }
-
-    #[test]
     fn scale_real_pyth_price() {
         // ATOM
         let uatom_price_in_usd = scale_pyth_price(1035200881u128, -8, 6u8, 6u8).unwrap();
